ddns

+++ Scripts/Handles.sh
@@ -84,8 +84,10 @@ if [ -f "$FW_FILE" ]; then
 fi
 
 #替换成最新版的golang
-cd $GITHUB_WORKSPACE/wrt/
-rm -rf feeds/packages/lang/golang
-git clone https://github.com/sbwml/packages_lang_golang -b 23.x feeds/packages/lang/golang
-cd $PKG_PATCH && echo "golang update done!"
+if [[ $WRT_REPO != *"lede"* ]]; then
+	cd $GITHUB_WORKSPACE/wrt/
+	rm -rf feeds/packages/lang/golang
+	git clone https://github.com/sbwml/packages_lang_golang -b 23.x feeds/packages/lang/golang
+	cd $PKG_PATCH && echo "golang update done!"
+fi
